@@ -32,32 +32,32 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  provider: ^6.0.3
-  intl: ^0.17.0
-  connectivity_plus: ^2.3.6+1
-  logger: ^1.1.0
-  shared_preferences: ^2.0.15
-  http: ^0.13.5
-  get_it: ^7.2.0
-  url_launcher: ^6.1.5
-  flutter_spinkit: ^5.1.0
+  cupertino_icons: ^1.0.6
+  provider: ^6.0.5
+  intl: ^0.18.1
+  connectivity_plus: ^4.0.2
+  logger: ^2.0.2+1
+  shared_preferences: ^2.2.1
+  http: ^1.1.0
+  get_it: ^7.6.4
+  url_launcher: ^6.1.14
+  flutter_spinkit: ^5.2.0
   enum_to_string: ^2.0.1
-  cached_network_image: ^3.2.1
-  flutter_svg: ^1.1.5
-  google_fonts: ^3.0.1
-  font_awesome_flutter: ^10.2.1
-  firebase_core: ^1.23.0
-  firebase_messaging: ^13.0.3
-  flutter_local_notifications: ^11.0.1
-  firebase_crashlytics: ^2.9.0
-  package_info_plus: ^1.0.6
+  cached_network_image: ^3.3.0
+  flutter_svg: ^2.0.7
+  google_fonts: ^6.1.0
+  font_awesome_flutter: ^10.5.0
+  firebase_core: ^2.16.0
+  firebase_messaging: ^14.6.8
+  flutter_local_notifications: ^15.1.1
+  firebase_crashlytics: ^3.3.6
+  package_info_plus: ^4.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  uuid: 3.0.6
+  uuid: ^4.1.0
   random_string: ^2.3.1
 
   # The "flutter_lints" package below contains a set of recommended lints to
@@ -65,7 +65,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^2.0.3
 
 # The following section is specific to Flutter packages.
 flutter:
